feat: commit

+++ src/main/java/com/example/skillnest/entity/OrderItem.java
@@ -30,5 +30,5 @@ public class OrderItem {
     Course course;
 
     @Column(nullable = false)
-    BigDecimal price;
+    Double price;
 }
